added blog update feature

+++ server/controllers/blogsController.js
@@ -4,7 +4,7 @@ const mongoose = require('mongoose')
 
 const getBlogs = async ( req, res) =>{
     try{
-        const response = await Blog.find({})
+        const response = await Blog.find({}).sort({createdAt: -1})
         res.status(200).json(response)
     }catch(error){
         console.log(error)
@@ -28,12 +28,23 @@ const getBlog = async (req, res) => {
 
 };
 
+const getUserBlogs = async (req, res) => {
+  const user_id = req.user._id;
+  try {
+    const response = await Blog.find({ user_id }).sort({ createdAt: -1 });
+    res.status(200).json(response);
+  } catch (error) {
+    console.log(error);
+  }
+};
+
 
 const newBlog = async (req, res) => {
 
-    const body = req.body
+    const {title, body, author } = req.body
+    const user_id = req.user._id
     try{
-        const response = await Blog.create(body);
+        const response = await Blog.create({title, body, author, user_id});
         res.status(200).json(response)
     } catch(error){
         res.status(400).json({"mssg" : "unable to create a blog"})
@@ -43,16 +54,25 @@ const newBlog = async (req, res) => {
 
 const deleteBlog = async (req, res) => {
   const { _id } = req.params;
+  const user_id = req.user._id;
 
   if (!mongoose.Types.ObjectId.isValid(_id)) {
     return res.status(404).json({ error: "Not a valid object id" });
   }
 
-  const blog = await Blog.findOneAndDelete({_id});
+  const blog = await Blog.findOne({_id});
 
   if (!blog) {
     return res.status(404).json({ error: "No document with this id exists" });
   }
+
+  if(blog.user_id != user_id){
+    return res.status(401).json({error: "You are not authorized to delete this document"})
+  } 
+
+  await Blog.findOneAndDelete({_id})
+
+  
   res.status(200).json(blog);
 };
 
@@ -75,4 +95,6 @@ const updateBlog = async (req, res) => {
 };
 
 
-module.exports = {getBlog, getBlogs, newBlog, deleteBlog, updateBlog}
+
+
+module.exports = {getBlog, getBlogs, newBlog, deleteBlog, updateBlog, getUserBlogs}
